ISSUE #687 moved some files around and created proper unit test holder
