+++ bouncer/src/repo/manipulator/drawingconverter/CMakeLists.txt
@@ -4,15 +4,4 @@
 #If you really need to overwrite this file, be aware that it will be overwritten if updateSources.py is executed.
 
 
-set(SOURCES
-	${SOURCES}
-	${CMAKE_CURRENT_SOURCE_DIR}/repo_drawing.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/repo_drawing_manager.cpp
-	CACHE STRING "SOURCES" FORCE)
-
-set(HEADERS
-	${HEADERS}
-	${CMAKE_CURRENT_SOURCE_DIR}/repo_drawing.h
-	${CMAKE_CURRENT_SOURCE_DIR}/repo_drawing_manager.h
-	CACHE STRING "HEADERS" FORCE)
-
+add_subdirectory(import)
