+++ bouncer/src/repo/manipulator/modelconvertor/import/odaHelper/file_processor_dwg.h
@@ -18,7 +18,7 @@
 #pragma once
 
 #include "../../../../core/model/bson/repo_node_mesh.h"
-#include "../../../../manipulator/modelconvertor/import/repo_drawing_import_manager.h"
+#include "../../../../manipulator/drawingconverter/import/repo_drawing_import_manager.h"
 
 #include <OdaCommon.h>
 #include <DbDatabase.h>
