+++ test/src/unit/repo_test_database_info.h
@@ -51,6 +51,8 @@ const static std::string ifcModel_InfiniteLoop = "infiniteLoop.ifc";
 const static std::string ifc4Model = "ifc4Test.ifc";
 const static std::string dgnModel = "sample.dgn";
 const static std::string dwgModel = "box.dwg";
+const static std::string dgnDrawing = "square.dgn";
+const static std::string dwgDrawing = "square.dwg";
 const static std::string dwgNestedBlocks = "nestedBlocks.dwg";
 const static std::string dwgPasswordProtected = "passwordProtected.dwg";
 const static std::string dxfModel = "box.dxf";
